+++ FE/src/utils/chart/drawUpperYAxis.ts
@@ -1,4 +1,4 @@
-import { Padding, StockChartUnit } from '../types.ts';
+import { Padding, StockChartUnit } from '../../types.ts';
 import { makeYLabels } from './makeLabels.ts';
 
 export const drawUpperYAxis = (
